nsupdate.info:ipv6 - fix detection method (use socket ipv6 detector)
web detector works well for ipv4 (we usually want the outside router IP there),
but not for ipv6, because often the src IP is privacy enhanced and thus temporary.

with the socket detector, we always get the static ipv6 address of the machine
(and even save doing a http(s) request because socket detector does not send a request).

considering this, I am not sure the web based ipv6 method ever makes sense, we maybe
can delay adding a profile for it until someone actually requests it.