move endless loop to the outside of the updater(s), so we can deal with all of them, fix daemonize
the granularity of the outer sleep is rather fine, so all (potentially different) sleep values of the updaters can be matched closely.
needs_check() is called every 10s now, but it is usually only a few lines of python (until it reaches the updater's sleeptime and triggers execution of more code).

also: only daemonize once